Dataset Description

This dataset contains information from a meta-analysis on multigenerational invasive exposure.

Usage

f2024

Format

A data frame with 1999 observations and 19 variables:

X

Integer. Observation number.

no_sd

Numeric. Standard deviation for 'no' group.

focal_species

Character. Species of interest.

treatment_level

Character. Treatment level description.

yes_n

Integer. Sample size for 'yes' group.

no_n

Integer. Sample size for 'no' group.

organ

Character. Organ type.

no_mean

Numeric. Mean value for 'no' group.

treatment

Character. Description of treatment.

invasive_species

Character. Invasive species name.

interaction_type

Character. Type of interaction observed.

yes_sd

Numeric. Standard deviation for 'yes' group.

yes_mean

Numeric. Mean value for 'yes' group.

study_id

Character. Study identifier.

native_functional_group

Character. Functional group of native species.

invasive_functional_group

Character. Functional group of invasive species.

latitude

Numeric. Latitude coordinate.

longitude

Numeric. Longitude coordinate.

collection_year

Integer. Year of data collection.

Details

Abstract: Invasive plant species devastate native plant communities, often with strong influences from belowground mechanisms. However, some native plants have responded with rapid trait change due to multigenerational interactions with invasives, including in root biomass allocation. Native plant root allocation responses in this context are poorly typified although invasive species are likely to have a pronounced influences belowground on native plants. Additionally, whether resource availability mediates the responses of invader-naïve and invader-experienced plants has yet to be summarized. To address these gaps, I meta-analyzed studies that evaluated root partitioning (as root mass fraction; RMF) from invader-experienced and invader-naïve native populations, in experiments in which native plants were grown alone or with the co-occurring invasive species at varying resource levels. Using 242 effect sizes on 17 native species, I found that differences in response to the presence or absence of multigenerational invasive exposure were only apparent when resource levels were applied. Specifically, when grown alone, invader-experienced plants had greater RMF under resource stress and lower RMF under high resource availability, and invader-naïve plants demonstrated the opposite pattern (p < 0.001). These results demonstrate invasive plant exposure produced adaptive responses to resources which are observed even without interactions with invasives. These results indicate a greater influence of invasive species belowground and call for more research assessing ecological and evolutionary root trait responses to coexistence with invasive species.

Examples

if (FALSE) {
  head(f2024)
}