Dataset Description

This dataset contains information on the water stress and competition meta-analysis

Usage

ff2019

Format

A data frame with 92 observations and 21 variables:

No.

Numeric. Observation number.

Study

Character. Name of the study.

Study_location

Character. Location of the study.

Target_species

Character. Species targeted in the study.

Target_life_history

Character. Life history of the target species.

Target_functional_group

Character. Functional group of the target species.

Study_type

Character. Type of the study (e.g., Outdoor, Indoor).

study_unit

Character. Unit of study (e.g., mesocosm, plot).

Water_trt

Character. Water treatment applied.

Comp_trt

Character. Component treatment applied (e.g., Shoot_Comp, Root_Comp).

Response_unit

Character. Unit of the response variable measured.

mean

Numeric. Mean value of the response variable.

se

Numeric. Standard error of the mean.

std

Numeric. Standard deviation.

N

Numeric. Number of observations.

Cntrl_Mean

Numeric. Mean value of the control group.

Cntrl_STD

Numeric. Standard deviation of the control group.

Cntrl_N

Numeric. Number of observations in the control group.

notes

Character. Additional notes or comments.

LRR

Numeric. Log response ratio.

LRR_var

Numeric. Variance of the log response ratio.

Source

https://doi.org/10.1371/journal.pone.0220674

Details

Abstract: Background: Competition is a critical process that shapes plant communities and interacts with environmental constraints. There are surprising knowledge gaps related to mechanisms that belie competitive processes, though important to natural communities and agricultural systems: the contribution of different plant parts on competitive outcomes and the effect of environmental constraints on these outcomes.Objective: Studies that partition competition into root-only and shoot-only interactions assess whether plant parts impose different competitive intensities using physical partitions and serve as an important way to fill knowledge gaps. Given predicted drought escalation due to climate change, we focused a systematic review–including a meta-analysis on the effects of water supply and competitive outcomes. Methods: We searched ISI Web of Science for peer-reviewed studies and found 2042 results. From which eleven suitable studies, five of which had extractable information of 80 effect sizes on 10 species to test these effects. We used a meta-analysis to compare the log response ratios (lnRR) on biomass for responses to competition between roots, shoots, and full plants at two water levels. Results: Water availability treatment and competition treatment (root-only, shoot-only, and full plant competition) significantly interacted to affect plant growth responses (p < 0.0001). Root-only and full plant competition are more intense in low water availability (-1.2 and -0.9 mean lnRR, respectively) conditions than shoot-only competition (-0.2 mean lnRR). However, shoot-only competition in high water availability was the most intense (— 0.78 mean lnRR) compared to root-only and full competition (-0.5 and 0.61 mean lnRR, respectively) showing the opposite pattern to low water availability. These results also show that the intensity of full competition is similar to root-only competition and that low water availability intensifies root competition while weakening shoot competition. Conclusions: The outcome that competition is most intense between roots at low water availability emphasizes the importance of root competition and these patterns of competition may shift in a changing climate, creating further urgency for further studies to fil knowledge gaps addressing issues of drought on plant interactions and communities.

Examples

if (FALSE) {
  head(ff2019)
}