Dataset Description

This dataset contains information from a plant competition experiment with Heterotheca villosa and Dieteria canescens.

Usage

ffk2024

Format

A data frame with 616 observations and 41 variables:

X

Integer. Observation number.

No.

Integer. Sample number.

Block_No

Integer. Block number.

Focal_species

Character. Species of interest.

Species_code

Character. Species code.

Population_code

Character. Population code.

Treatment

Character. Type of treatment.

Comp_trt

Character. Competitor treatment.

Competitor

Character. Competitor species.

Time_in_Exp

Integer. Time in the experiment.

initial_nonfocal_leaf_count

Integer. Initial leaf count for non-focal species.

initial_focal_leaf_count

Integer. Initial leaf count for focal species.

nonfocal_root_length

Numeric. Root length for non-focal species.

nonfocal_shoot_height

Numeric. Shoot height for non-focal species.

nonfocal_root_mass

Numeric. Root mass for non-focal species.

nonfocal_shoot_mass

Numeric. Shoot mass for non-focal species.

focal_root_length

Numeric. Root length for focal species.

focal_shoot_height

Numeric. Shoot height for focal species.

focal_root_mass

Numeric. Root mass for focal species.

focal_shoot_mass

Numeric. Shoot mass for focal species.

Trt

Character. Treatment details.

focal_species

Character. Focal species name.

Type

Character. Type of plant.

nonfocal_leaf_count

Integer. Leaf count for non-focal species.

focal_leaf_count

Integer. Leaf count for focal species.

focal_full_mass

Numeric. Full mass for focal species.

nonfocal_full_mass

Numeric. Full mass for non-focal species.

focal_rmf

Numeric. Root mass fraction for focal species.

nonfocal_rmf

Numeric. Root mass fraction for non-focal species.

rmf_diff

Numeric. Difference in root mass fraction.

shoot_height_diff

Numeric. Difference in shoot height.

root_length_diff

Numeric. Difference in root length.

leaf_count_diff

Integer. Difference in leaf count.

pot_mass

Numeric. Pot mass.

abs_rmf_diff

Numeric. Absolute difference in root mass fraction.

abs_root_length_diff

Numeric. Absolute difference in root length.

abs_shoot_height_diff

Numeric. Absolute difference in shoot height.

final_leaf_diff

Integer. Final leaf count difference.

abs_final_leaf_diff

Integer. Absolute final leaf count difference.

initial_leaf_diff

Integer. Initial leaf count difference.

abs_initial_leaf_diff

Integer. Absolute initial leaf count difference.

Details

Abstract: Evaluating restoration practices such as admixture provenancing, or the mixing of multiple source populations for greater intraspecific variation in plant restorations, helps to understand the utility of the practice in different scenarios. For admixture provenancing, we do not know whether just any population mixture will deliver positive outcomes in mixed scenarios. To understand how populations and species identity influence plant performance and trait outcomes in mixtures, we assessed trait variation among populations of two forb species native to the western US (Dieteria canescens and Heterotheca villosa) following three interaction treatments (single-population monoculture, two-population mixture, and two-species mixture) with Dieteria canescens as a focal plant in a greenhouse common environment. Plants were harvested after seven weeks of growth and plant height, leaf count, root length, and root mass fraction traits were measured, as well as whole plant biomass. We found that four of six population pairings led to more positive interactions compared to single population monocultures. Neighbor trait differences in shoot height and root length best predicted plant growth following interactions, while increasing neighbor trait differences resulted in positive growth outcomes for all but one trait by treatment interaction. In this single case, plant growth decreased with increasing differences in root mass fraction for the two-species mixture treatment, surprisingly. These findings show nuance in the outcomes of interactions within and between populations and species: they are highly dependent on population identity and traits measured, which are aspects that must be considered when evaluating restoration materials.

Examples

if (FALSE) {
  head(ffk2024)
}